+++ src/BlogPost.ts
@@ -5,7 +5,6 @@ import { writeFile, readFile, readFileIfExists } from "./helpers/io"
 import { copy } from "fs-extra"
 import { parse } from "path"
 const frontmatter = require("front-matter")
-const marked = require("marked")
 import markdownIt from "markdown-it"
 
 const md = markdownIt({
@@ -68,10 +67,8 @@ export default class BlogPost implements IResource {
     async write(path) {
         // Copy all dependencies
         // TODO: do this better using the new v0.2.0 structure
-        // console.log("COPY", parse(this.path).dir)
         await copy(parse(this.path).dir, parse(path).dir, {
             filter: (src: string, dest: string): boolean => {
-                // console.log(src, ",", this.path)
                 return src !== this.path
             }
         })