+++ src/Builder.ts
@@ -52,7 +52,6 @@ export default class Builder {
                 })
                 post.write(buildPath)
             } else {
-                console.log("draft post")
             }
         })
     }
@@ -76,9 +75,14 @@ export default class Builder {
         let pageBuildDir = join(this.dirBuild, page.name)
         makeDir(pageBuildDir)
         let pageBuildFile = join(pageBuildDir, "index.html")
-        page.render({
-            meta: this.nconfig.meta
-        })
+        try {
+            page.render({
+                meta: this.nconfig.meta
+            })
+        } catch (err) {
+            this.cli.warn(err)
+        }
+
         await page.write(pageBuildFile)
         return pageBuildFile
     }
@@ -130,9 +134,13 @@ export default class Builder {
             let postPath = join("build", page.name)
             makeDir(postPath)
             let buildPath = join(postPath, "index.html")
-            page.render({
-                meta: this.nconfig.meta
-            })
+            try {
+                page.render({
+                    meta: this.nconfig.meta
+                })
+            } catch (err) {
+                this.cli.error(err)
+            }
             page.write(buildPath)
         })
 
@@ -218,7 +226,6 @@ export default class Builder {
                         } catch (err) {
                             this.cli.error(err)
                         }
-                        console.log(blogPost.attributes)
                         // Ignore draft posts
                         if (!blogPost.attributes.draft) {
                             return blogPost